requirements.txt
langchain
langchain-google-genai
pypdf
python-dotenv
langchain_community
pydantic
langchain_core